feat: CldVideoPlayer support for Next.js 13 App Router (#265)
# Description

With the App Router, we no longer have access to the <Head component.
CldVideoPlayer was utilizing this component in order to load a
stylesheet, preventing the need to include it from the project, but
without <Head, the player breaks.

Given this is no longer an option, this includes the CSS and associated
files on build so that they're able to be imported from within the App
Router project, allowing the stylesheet to be imported for a working
player.

This is currently downloading the assets from unpkg, where the player is
officially hosted, though future iterations may include these files from
the cloudinary-video-player node_module, though thats currently on hold
for performance reasons. See #181

This also deprecates the `version` prop as we're no longer able to
reliably grab assets based on the configured version (or at all for that
matter). In future versions using the node module, the assets will be
synced based on the installed dependency via next-cloudinary without
configuration available.

## Issue Ticket Number

Fixes #248 

<!-- Specify above which issue this fixes by referencing the issue
number (`#<ISSUE_NUMBER>`) or issue URL. -->
<!-- Example: Fixes
https://github.com/colbyfayock/next-cloudinary/issues/<ISSUE_NUMBER> -->

## Type of change

<!-- Please select all options that are applicable. -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Fix or improve the documentation
- [ ] This change requires a documentation update


# Checklist

<!-- These must all be followed and checked. -->

- [ ] I have followed the contributing guidelines of this project as
mentioned in [CONTRIBUTING.md](/CONTRIBUTING.md)
- [ ] I have created an
[issue](https://github.com/colbyfayock/next-cloudinary/issues) ticket
for this PR
- [ ] I have checked to ensure there aren't other open [Pull
Requests](https://github.com/colbyfayock/next-cloudinary/pulls) for the
same update/change?
- [ ] I have performed a self-review of my own code
- [ ] I have run tests locally to ensure they all pass
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes needed to the documentation