docs: add aspect ratio docs for CldImage component (#544)
# Description

<!-- Include a summary of the change made and also list the dependencies
that are required if any -->
This PR adds documentation for the `aspectRatio` prop of the `CldImage`
component, highlighting the constraints and cases where `aspectRatio`
prop is applicable and will work properly.

The constraint is that `width` and `height` props are required by NextJS
and `aspectRatio` prop does not function properly when they are set. So,
for proper functioning of `aspectRatio`, the `fill` prop must also be
present on the component, and `height` and `width` props must be absent.

More details
[here](https://github.com/cloudinary-community/next-cloudinary/issues/475)

## Issue Ticket Number

Fixes #475 

<!-- Specify above which issue this fixes by referencing the issue
number (`#<ISSUE_NUMBER>`) or issue URL. -->
<!-- Example: Fixes
https://github.com/cloudinary-community/next-cloudinary/issues/<ISSUE_NUMBER>
-->

## Type of change

<!-- Please select all options that are applicable. -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Fix or improve the documentation
- [x] This change requires a documentation update


# Checklist

<!-- These must all be followed and checked. -->

- [x] I have followed the contributing guidelines of this project as
mentioned in [CONTRIBUTING.md](/CONTRIBUTING.md)
- [x] I have created an
[issue](https://github.com/cloudinary-community/next-cloudinary/issues)
ticket for this PR
- [x] I have checked to ensure there aren't other open [Pull
Requests](https://github.com/cloudinary-community/next-cloudinary/pulls)
for the same update/change?
- [x] I have performed a self-review of my own code
- [x] I have run tests locally to ensure they all pass
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes needed to the documentation